Changes in seed to main function for add_carbon (#223)
Usage of set_seed only by clm commands - individual functions do not support a seed anymore, but rely on the caller on setting it if reproducibility is desired. Using set_seed in all commands where it was used previously in the nested function. Passing seed from the snakemake step whenever we can.

---------

Co-authored-by: Vineet Bansal <vineetbansal@protonmail.com>